BACKLOG-9625 Reorganize code
