+++ src/javascript/components/contentEditor/EditvanityUrlsDialog/EditVanityUrlsDialog.scss
