+++ src/javascript/components/contentEditor/EditvanityUrlsDialog/index.js
@@ -1,0 +1,3 @@
+import EditVanityUrlsDialog from './EditVanityUrlsDialog';
+
+export default EditVanityUrlsDialog;
