+++ src/javascript/components/contentEditor/SiteSettingsSeo/SiteSettingsSeoEntry.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import SiteSettingsSeoCard from './SiteSettingsSeoApp';
-import SiteSettingsSeoWrapper from '../SiteSettingsSeoWrapper';
+import SiteSettingsSeoWrapper from '../../SiteSettingsSeoWrapper';
 
 const SiteSettingsSeoCardEntry = props => <SiteSettingsSeoWrapper component={SiteSettingsSeoCard} {...props}/>;
