+++ src/javascript/components/contentEditor/SiteSettingsSeo/index.js
@@ -1,0 +1,3 @@
+import SiteSettingsSeoEntry from './SiteSettingsSeoEntry';
+
+export default SiteSettingsSeoEntry;
