+++ src/javascript/components/contentEditor/VanityUrlTableData/VanityUrlTableData.jsx
