+++ src/javascript/components/contentEditor/VanityUrlTableData/index.js
@@ -1,0 +1,3 @@
+import VanityUrlTableData from './VanityUrlTableData';
+
+export default VanityUrlTableData;
