+++ src/javascript/components/contentEditor/actions/index.js
@@ -1,0 +1,5 @@
+import VanityAction from './VanityAction';
+
+export {
+    VanityAction
+}
