+++ src/main/resources/javascript/locales/de.json
@@ -110,6 +110,6 @@
         "of": "von"
     },
     "title": "Vanity URLs",
-    "manage": "Manage Vanity URLs"
+    "manage": "Vanity URLs verwalten"
   }
 }
